+++ python/case_uco/__init__.py
@@ -17,4 +17,4 @@ __all__ = [
     "InvalidSplitSizeError",
     "clear_class_registry_cache",
 ]
-__version__ = "1.22.2"
+__version__ = "1.22.3"